♻️ Replace dark mode inline JS with Stimulus controller
Introduce a dark-mode Stimulus controller that manages theme toggling,
icon state, aria-pressed, localStorage persistence, and system
preference listening.

Replace id-based dark mode toggle buttons and the mobile onclick handler
with declarative data-controller/data-action attributes. Remove ~80
lines of dark mode JS from app.js (initializeDarkMode + window.toggleDarkMode).

The inline IIFE in base.html.twig and error templates stays as-is to
prevent FOUC — it runs before any JS bundle loads.

Part of #1045
